feat(dal) Update update_by_id_v1 to account for read/write tenancy split
update_by_id_v1 was operating under the old behavior where the read &
write tenancies were identical. Unfortunately, this meant that if you
tried to update a row that was in the universal tenancy, you were unable
to if you were not also in the universal tenancy.

Now, we also look in the universal tenancy when we didn't initially find
a version to update where the tenancy is exactly the write tenancy. We
then copy that record into our write tenancy.