chore(dal) Remove SchemaId from InternalProvider
Because Props are unique across SchemaVariants, and SchemaVariants can
only belong to one Schema, we don't need to specify which Schema an
InternalProvider is for.

This commit also ensures that we always create the Prototype for an
InternalProvider as the InternalProvider only context, which brings it
in line with how we handle Props (which have a Prototype/Value at the
Prop only context when they are created).