feat(dal): add post fix intelligence test and three regression tests
The post fix intelligence test for resource is currently failing due to
a known bug. This test is being added with the "ignore" macro so that
others can use it.

The three regression tests should currently pass and provide a baseline
for testing in the future. They can (and likely should) be expanded to
account for multiple scenarios.

Finally, fix clippy lints due to Rust 1.65's release.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>