bug(dal) Only PropKind::Object is re-usable across AttributeContext; all others must be created new
Previously, this check was only requiring Array/Map to exist in the
exact AttributeContext that we were interested in, but this is
incorrect, as it forgot about all of the scalar types. Across all
PropKind, Object is the only one that can be re-used across
AttributeContext. Object's special treatment is also likely to go away
at some point, so it can use the IndexMap for determining child Prop
display order.