fix(sdf): update_property_editor_value should lookup value by context first
When making the first update to a value, the blur event might be
triggered twice quickly in a row. We should look into synchronizing
those blur updates so that we don't have overlapping ones going on at
once, but in this case, the updates did not actually overlap. Instead,
the update_for_context method was being called twice in a row with the
attribute value of the "base" prop, since a value did not yet exist for
the prop + component context. This caused two attribute values to be
created. The validation would be run for the second one, but the
property editor would get the attribute value id of the first one, and
so we wouldn't display validations for the value. This ensures we
lookup the existing attribute value by the provided context, and
update *that* attribute value if it is a different one than the one sent
in the request.