fix(web): skip updates if value id is out of date
This is a stop gap before implementing better update synchronization.

When updating a value that does not yet exist in the component context,
we only have the value id for the schema variant default value. If
blur is triggered fast enough, the update can fire before the vue
reactive tree has been able to update the value id from the
get_property_editor_values request. This can be the case *even if* the
request has completed and is not currently in flight.

We should just not execute the update in this case. We should also
debounce the blur, probably. It would also be ideal if the update
operation was idempotent, or close to it, when updating the value for a
schema variant level attribute but passing in a component level context.

There are some other validation bugs I'm tinkering on as a follow up.