feat(dal,sdf): exec_func endpoint supports validation funcs
Makes it possible to just execute one validation func on a component (a
single validation for a single prop).

Reworks the validation_resolvers table slightly for clarity. Adds a
uniqueness constraint to the table so that there can only ever be one
validation resolver per attribute_value_id and func_id. The
func_binding/return_value caching mechanisms ensure we don't have to
re-execute identical functions, but we need to be able to modify a
function for the same value and same validation result but return a
different message without creating another resolver.