feat(dal) Make set_belongs_to_v1 update pre-existing relation records
Previously, you had to manually `unset_belongs_to_v1` before you could
change the `belongs_to_id` of an object, which caused issues with
merging changesets, and was generally unintuitive. Now we use the
existing `find_by_attr_v1` logic to see if there is already a relation
record for the `object_id` we're interested in and, if there is, use
`update_by_id_v1` to update that record according to our "normal"
changeset logic, instead of always inserting a new record into the
changeset.