feat(dal): add child internal providers to the dependency graph
Currently we only look up towards the root for internal providers when
constructing the dependency graph. However, we need to also look down
towards the leaves so that we include the child props/internal providers
in the update graph. Otherwise functions that take those child internal
providers as an argument will either not get triggered, or will get
triggered with the old value since we won't have emitted the updated
value to the implicit internal providers.