feat(dal): expose `ActorView` for all `WsEvent` messages
This augmented type has a `label` field which contains a set value for
system actors and the user's display name for users. It comes at the
cost of a user lookup by id, but as consumers of the API (i.e. the
frontend) often want to display human-relatable output, this ships such
a value rather than having to fallback on displaying raw identifiers or
performing secondary lookups.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>