bug(dal): Let records deleted within a change set "mask" the HEAD version
Previously, if a record existed in a change set with the
`visibility_deleted_at` set, while there was a version of it on `HEAD`
that did not have `visibility_deleted_at` set, the change set would
never treat the record as deleted, even though merging the change set to
`HEAD` would "delete" the record.

Now, when we're looking at records in the `<table>_v1` functions, we do
a self-join to be able to see if the `HEAD` version of a record also has
one specific to the change set we're interested in. If the change set
version has `visibility_deleted_at` set, then we don't consider the
`HEAD` version for the purposes of the `DISTINCT ON`, effectively also
treating it as deleted.