chore(dal): avoid failing when the name is not available
Name is set when the component is created, but most places get name
from components_with_attributes which uses the root's internal provider,
which only gets propagated by the DependentValuesUpdate job, so even
though the name might be there, the value won't be found.

So instead of failing when the name is not found we use a default name.

We also avoid sending the change set written event through the websocket
in node creation because although it was written it's incomplete, so it
will help the frontend to not reload the partially saved data.
DependentValuesUpdate will send its own ChangeSetWritten. It doesn't fix
all of the possible races, but it's an improvement in the regular workflow.