feat: create initial packaging system
This is a first implementation of a packaging system for SI which will
allow end users to export authored content from the platform (such as
Schemas, Functions, etc.) and import the same content into new
deployments of the platform.

Code Organization
=================

There are 3 parts to this system:

An Object Tree
--------------

Most of the implementation for object trees are in the `lib/object-tree`
Rust crate. This is a way to represent a tree structure over an
arbitrary node type that is cryptographically hashed and as such is a
low-level, generic crate that doesn't contain any SI-specific business
logic. The crate provides traits that when implemented allow a consumer
of the crate to convert types into a hashed tree. For more details about
the tree structure and some supporting reading, see the module comment
for the crate at the top of `lib/object-tree/src/lib.rs`.

Additionally this crate provides support to serialize these object trees
to a file system format--currently into an output directory and to a Tar
file as well as supporting a de-serializing from these file system
formats back into an object tree.

Packaging-Specific Types and Object Tree Trait Implementations
--------------------------------------------------------------

The implementation for these types and trait implementations are in the
`lib/si-pkg` Rust crate. These types represent packaging primitives that
ultimately get inserted into an object tree as part of the packaging
hierarchy. For example, there is a `struct Prop` which importantly is
not the same as the `dal::Prop` type. This separation allows us to
evolve the packaging subsystem more independently from the DAL business
logic code. There are also likely types in this crate that won't have a
1:1 relationship with a DAL type--think of this crate's types as having
more of a serialize/de-serialize set of responsibilities.

Additionally there are some more types and trait implementations which
allow the outer packaging types to be converted into and out of an
object tree.

DAL Integration
---------------

Guess where this implementation lives? `lib/dal` you say? Right! This is
where we consume the si-pkg types and object tree implementation and use
them to either create/add new DAL objects (such as Schemas,
SchemaVariants, Props, Providers, etc.) or to export existing DAL
objects from the current system.

Current Functionality
=====================

This change introduces the bare minimum functionality while providing
the basis for extending this system is a logical way. As a result,
importing and exporting a tree of props for a schema variant is the only
supported content of a package. That is, from the "domain" prop down.
Future work will add support for handling a schema, schema variant,
functions, multiple schemas, etc.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>