feat(web,dal,sdf): Add ability to create delete confirmations
When a component is deleted in a changeset, a new flag on the components
table `needs_destroyed` is set to true. When the changeset is merged,
the delete confirmation for a component (if there is one) will run
based on there being a `visibility_deleted_at` timestamp AND the
`needs_destroyed` field being true.

When the fix is applied (i.e. the resource is deleted), the value of
`needs_destroyed` is set to false so the fix flow understands that the
component is dealt with

Co-authored-by: Paulo Cabral Sanz <paulo@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Theo Ephraim <theo@systeminit.com>