chore(sdf): refactor scenario tests to be children of the module
Refactor scenario tests to be children of the module for a few reasons.
Some of which include the following:

- The scenario tests were getting too big to be contained in the same
  module (now each test is in its own submodule)
- The harness struct was a public struct, but it should be private and
  only available to scenario tests (now it is)

We also renamed the "model_and_fix_flow" scenario test to
"model_and_fix_flow_whiskers" to imply that it replicates the full
flow performed in one of our usability tests.

Signed-off-by: Nick Gerace <nick@systeminit.com>