fix(dal): Properly select return data from delete_by_id_v1 SQL function
Previously, the way the select was being done, it was expecting the
return of the function to be named `deleted_at`, instead of `updated_at`
(which is what the function had its out variable named). The select
itself was failing, because the resulting anonymous table created by
using the function as a table function did not have a `deleted_at`
column.

By using the function as a scalar function, instead of as a table
function, we no longer need to be concerned with what the out variable
is named, only that it returns a single, scalar value.