fix(dal) Do not ignore result return value in test
This test was hiding an issue during the execution of the function,
because it was explicitly ignoring the result returned by the function.
This was causing the test to fail with a panic later on, when it would
try to use the PG transaction that was in a rollback state, and was no
longer usable.