feat(dal,si-pkg,object-tree): add pkg metadata & schemas/variants
This change adds top-level package metadata such as name, created by,
version, etc. and the packaging format supports adding multiple schemas
with multiple schema variants. In other words the following is
package-able (note we're not using YAML in the impl, but this shows the
shape of data):

```
name: Docker Stuffs
version: 0.1.0
schemas:
  - schema: Docker Image
    - variant: v0
      - prop: domain
          - prop: image
  - schema: Docker Hub Credential
    - variant: v0
      - prop: domain
          - prop: secret
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>