fix(ci): restore conditional Docker image building
This change fixes a bug bug introduced when migrating to the updated
GitHub Actions [environment files] format from the soon-to-be-deprecated
`set-state`/`set-output` commands.

The format for each line of output is `{name}={value}` (for example
`bin--sdf=true`) and we were producing lines like
`name=bin--sdf::true`. As these lines weren't parsable as environment
variables, none of the component environment variables were being set,
and so the `build-docker-images` CI workflow determined there were no
Docker images to be built (i.e. the `if` check in each component job
evaluated to `false` every time).

As an additional cleanup, the entire block of these environment
variables are appended to the GitHub Actions environment file with on
append action to prevent any future copy/paste errors.

[environment files]: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#environment-files

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>