fix(ci): run all Rust targets only when necessary
This change updates the intention of the `mk/test-changed.sh` script to
run all Rust tests only when a global config file such as
`Cargo.toml`/`Cargo.lock`/etc. has been modified in a Git change set.

The prior implementation caused some PRs to run all tests for trivial
changes (like *.md files in the root) or to die when a component was
living under `component/` (which don't have a `Makefile`).

Now, when the following files are present in the list of changed files,
all Rust `test//*` Make tasks will be run:

- `Cargo.lock`
- `Cargo.toml`
- `rust-toolchain`

Additionally, when the following files are present in the list of
changed files, all Rust `check//*` Make tasks will be run:

- `rustfmt.toml`
- `clippy.toml`

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>