feat(dal,sdf): track installed assets when installing pkg
Assets installed by a package are added to the installed_pkg_assets
table during install. If an asset is already installed, we do not
install the asset but we still make a record of it as if it had been
installed by the package. This way we can prevent removal of an asset if
it is shared by more than one package. Using the hash, we also prevent
installing the same package twice.