bug(pg): Do not consume an entire core waiting for PG to finish booting
Turns out that the timeout on `pg_isready` doesn't apply if there is
nothing listening on the TCP port at all, and the command exits
immediately. Introducing the `sleep 1` means that we don't spin as fast
as we can, tying up a CPU core checking if PG is up yet.