chore(db): Make attribute_value_remove_value_and_children_v1 iterative
Previously, this function was written recursively, where it would call
itself to do a depth-first removal of the tree of AttributeValues.
Unfortunately, this could sometimes overflow the stack. Now, we gather
up the AttributeValues of the tree using a recursive query (which PG
executes iteratively), and loop over those results to remove the tree of
AttributeValues, which avoids stack usage that is dependent on the
size/depth of the tree of AttributeValues.