fix(dal): fix delete handler and ec2 delete function
Rename delete workflow functions to have the appropriate handler.
Use the correct InstanceId when deleting the ec2 instance.

Update resources of deleted components. This means if someone deletes
the real life resource of a deleted component, the component will go
away instead of waiting for the delete workflow to execute (which will
fail, as the thing is not there anymore)