chore(pg) Temporarily remove pganalyze from the PG container setup
We don't want to be running pganalyze whenever LOCAL_PG=true is used in
development, and we don't want to have the collector API key checked in
to the repo. For now, this removes the addition to have the pganalyze
collector submit stats to the API at all (both in dev and prod), until
we set up the production container to get the API key without having it
checked in to the repo.