fix(dal): don't run resource sync on components being fixed
Note: The only open problem here is that we re-run confirmations for the
component after the individual fix runs, that's why the recommendation
disappears from the panel before the fix batches finish. Because we have
to propagate resource updates so things like SecurityGroupId can be used
by subsequent fixes. But this means we have to propagate confirmations.
There is no way to exclude a attribute value from a specific propagation
of one of their parents.

Since dependent values update doesn't get inlined anymore we don't need
to avoid enqueueing it, we just need to ensure it blocks other tasks if
needed.

We don't need to schedule all confirmations runs in resource syncing, as
confirmations will be automatically run because of their dependency on
/root/resource.

Resource sync ignores components in an active fix batch.