fix(dal): don't show qualifications that haven't run
If a qualification exists on the qualification map but the function has
not yet run we show the wrong thing in list_qualifications. If a
qualification has not run, we should not show it yet. We can detect this
by checking the func binding return value's func id and comparing it to
the func id on the attribute prototype.

This fixes the "si:setObject" qualification displaying.

To see a newly authored qualification it must be executed from the func
details panel or an attribute on the component's domain must be updated
so that it is run via a dependent values update.