fix(si-data-pg) add Drop impl to pg client to ensure healthy reuse
This change adds some addtional safetey around a Postgres
connection/client just before the connection is being returned to the
Deadpool connection pool.

Our current implementation in `si-data-pg` wraps both the pg pool and
the client/connections which allows us to insert checks and/or logic
around the time when the objects are being torn down (i.e. the Drop impl
is being called).

This change primarily adds a `Drop` impl on the `InstrumentedClient`
type which takes ownership of the underlying Deadpool `Object` and
allows us to perform a query check before deciding whether or not to
return it to the pool. Our query check is to ask for the current
transaction ID of 2 independant queries which should return 2 different
IDs. If both IDs are the same, then the connection still has a
transaction in flight and in scope for those queries. In this situation
we are now going to explicitly remove this connection from the pool so
that the connection can be torn down and a new connection will be
created back in the pool (when needed).

Note
----

Remember that the original sign of problems was that we had clients
connected to the database server that were idle and did not have a query
in flight but does have an open transaction. After investigating the
upstream Deadpool and Postgres crates, we found a potential issue in
`tokio-postgres`. Namely, when a `Transaction` is issuing a `ROLLBACK`
because it is dropped before `COMMIT` or `ROLLBACK` has been explicited
called, the underlying connection does not appear to be polled to flush
the message back to the server. Chances are this can be future work on
the upstream to address this, however the change here is additional
sanity and safety which wouldn't interfere with an upstream change.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>