fix(dal): avoid duplicating rows with the same id in the same change-set
Makes visibility_tenancy UNIQUE INDEXES more strict, only allowing one
row with each id in each change-set, instead of allowing one deleted and
one live. This would block the change-set-apply of the broken change-set
instead of making them duplicate on head too.

Makes change_set apply more strict by checking for the tenancy (not
strictly needed as the change_set_pk should be unique on that call, but
it's more correct.

And the most important part of this change, now we prevent rows from
being incorrectly duplicated by `update_by_id_v1`. Before if a element
was deleted in the change-set, updating it without seeing deleted stuff
would duplicated it. As the update would not find it, hitting the insert
branch of update_by_id_v1, which would see that the element is live on
head and clone it to the change-set. Effectively ending up with two rows
with the same id in the same change-set, only differing by
visibility_deleted_at.

Now it will throw an exception, as an update is happening to something
that is not visible.

This won't fix the root cause that was deleting an
attribute_value_remove_value_and_children_v1 that was already deleted
without the visibility_deleted_at set in the context. We need to figure
that out.

Co-Authored-By: Zack Hamm <zack@systeminit.com>