chore: use more txns on builtin migration
This PR adds a bunch more transactions to the builtin
migrations.

We saw behavior where the system would come up, migrate
builtins, and then spend a few minutes cranking through the dependent
values updates. That's because the updates themselves are helpfully
batched up with the transaction - and we used one giant transaction for
all the work.

We now create new context and commit the transactions as we go, which
allows the dependent value update jobs to start happening in real time,
which makes the total time for the system to boot and quiesce much
shorter (maybe 30 seconds to a minute after finishing migration, rather
than 5+ minutes.)

Signed-off-by: Adam Jacob <adam@systeminit.com>