feat: posthog tracking for web and sdf
This PR adds posthog tracking for web and sdf. For sdf, we needed a new
library for tracking (the posthog rust library is way too simple). This
PR adds 'si-posthog-rs', which uses an mpsc channel to allow lots of
senders and a single posthog sender task. The result is that sending a
task should not block handlers returning, and if there are errors in
sending, we just print a trace statement.

We should try and avoid using `si-posthog-rs` within the dal - ideally,
we should be able to only track things in handlers in sdf.

Signed-off-by: Adam Jacob <adam@systeminit.com>