fix(dal): add Security Group Description regex validation
I noticed we do not check the length or validity of the Security Group
description. The create dry run qualification does not seem to check it
either, so it won't be noticed until actual security group creation.
This is a perfect case for a custom validation function which tests the
description against a regex.