fix(dal): avoid lingering transactions in dependent values update
We were keeping a transaction open when waiting for attribute values to
process because tests can't just commit and create another one.

This open transaction may exhaust deadpool's connections limit, making
things hang waiting for a connection to be available. It's less likely
to cause a dead-lock than the previous similar bug we fixed. But it can
still make things slow and at worse hang with one job getting a
transaction to wait for council and the other, that got a message from
council that has just finished processing the last one trying to clone
with new transaction.

This touches complicated parts of the system so I'm not eager to merge
this so close to the deadline, but it may be important enough. We should
review it thoroughly and play with it before merging it though.