fix(dal,web): fix JsAttribute func reversion and is_revertible status
If a prototype is on a changeset for a Prop context and a corresponding
prototype for the exact same context exists on head, it's safe to
delete! We need this to hard delete changeset prototypes for JsAttribute
functions via the "revert" endpoint. Also fixes the reactivity on save
of the isRevertible/is_revertible flag.