feat(dal,web): remove recommendation "is_runnable" and "status"
The following changes clear some cruft when working with
recommendations and tracking their associated fix states. This commit
does not fix a known bug where recommendations cannot be re-used if a
resource sync triggers a recommendation that's already been ran for a
given component. However, this commit does remove some of the noise that
may help in solving that bug.

Contents:
- Remove "is_runnable" and "status" fields for recommendations
- Use "last_fix" and "has_running_fix" fields to determine the state of a
  given recommendation (while still ensuring that two concerns are
  separate: the ability to run/re-run the recommendation and the status
  of the last time the recommendation was ran)

Signed-off-by: Nick Gerace <nick@systeminit.com>