fix(dal): updated `StatusUpdater`'s API to be infallible
This change updates `StatusUpdater`'s public API to not return `Result`
types, but rather deal with errors internally and to log errors as it's
being used.

Prior to this change an error in any of StatusUpdater's methods would
cause the caller to early-return with an error--and the only code that
uses StatusUpdater is the DependentValuesUpdate job code. In other
words, we don't want the updating logic to error the job's business
logic code.

Note: future work needs to account for what happens when a StatusUpdater
can't be properly initialized or when an error is encountered when it
runs its functions. Current we log an error and continue on but the
implication is that no NATS message may be transmitted which results in
`WsEvents` sent to connected frontends via the updating WebSocket.

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>