chore(ci,cyclone,telemetry): set Cyclone server default logging to `warn`
This change changes the default logging verbosity of Cyclone server
instance to `warn` rather than the previous default of `info`. Prior to
this change Cyclone server logging output would be seen in Veritech's
console output (when running Veritech in a terminal) and in integration
test output for dal and sdf tests. Note that `warn` and `error` level
logging will still be written out.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>