chore(dal-test): run each dal/sdf test with its own database copy
This change alters the database setup when running a DAL integration
test (using our `#[dal_test]`/#[test] macro) or SDF service integration
test (using the new `#[sdf_test]` macro).

Prior to this change, every test used the same migrated and prepared
database (i.e. `si_test`) and heavily relied on transactions and
rollbacks to get a high degree of test isolation. As our system has
matured and expanded into other subsystems such as Pinga, Council, etc.,
it's become more clear that we cannot rely on this tranastion/rollback
trick in our tests while at the same time ensuring the code execution
path is the same in test and in production. Our biggest problem right
now is that the code path for a DAL integration test is sufficiently
different from the code path used when the full system is running,
meaning that it gets harder to trust our integration tests to "act the
same" when a real user is driving our system.

The trick used here is to create a new database copy for each test,
using the migrated database as a template. As before, the first test to
"win" will perform the initial database migrations and run builtin code.
With this change, a version of this SQL is executed before each test:

```sql
CREATE DATABASE si_test_975da WITH TEMPLATE si_test OWNER si;
```

Then the `PgPool` is created to use this database and the pool is passed
through the rest of the test setup code for use by the test function.

The other change is that when the first test is about to prep the
`si_test` database for migrations, we attempt to drop all `si_test_*`
databases that were created in the last test run.

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>