fix(dal,web): fix status updates when loading recommendations/running fixes
Adds an "unstarted" state to individual fixes and the fix batch,
corresponding to the unstarted recommendation state and starts using it
again in the fix store. Returns incomplete fixes in fix/list with the
unstarted state and a null resource to get live updates working and
fixes a bug where we "finished" the batch on the frontend and loaded
only finished fixes while the batch was still running.

Also resolves ENG-1276 by adding the global status update bar to the fix
screen. It shows if recommendations are not running.

Co-Authored-By: theo@systeminit.com