chore(si-posthog): refactor client creation API to use a `PosthogConfig`
This change adds a `PosthogConfig` which allows us to add more PostHog
configuration into the client as we go. It also enabled a nice default
behavior and a builder-style creation of the client which benefits from
the defaults.

The biggest behavior change is that now that there's a config to
determine whether the client is enabled, we can disable it on creation
rather than post-creation over the internal channel.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>