feat(sdf,dal-test,sdf-server,si-test-macros): introduce `#[sdf_test]`
This change adds a new procedural function test macro called
`#[sdf_test]` which can be used in SDF service tests. The old code that
relies on `test_setup!` is not entirely removed.

As there is a lot of re-usable logic between the `#[dal_test]` and
`#[sdf_test]` implementations, a lot of this code was extracted and
threaded through an internal trait so that both impls could benefit (the
idea here is that for any common additions/bug fixes we should only need
to update one code site and not 2).

Additionally to help with test-time setup, the
`sdf_server::build_service()` function signature was updated to take a
`ServicesContext` which is a minor change in the `bin/sdf` server code
and a massive help for setting up an SDF service test.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>