chore(pg) Increase minimum query execution time to trigger autoexplain
Having a minimum query time of 250 ms is very agressive, and results in
triggering a _lot_ of autoexplains during normal usage, and during test
runs. Bumping the minimum query time to 1.5 seconds will make the
postgres container logs a bit less noisy. We'll still get useful
autoexplain output for the worst offenders, and can decrease the minimum
execution time to trigger autoexplain again in the future.